Refactor: Extract security functions from interpreter.js
This commit extracts the security and permission-related methods from the `RexxInterpreter` class in `core/src/interpreter.js` into a new, separate module `core/src/interpreter-security.js`.

This is a continuation of the effort to modularize the interpreter, making it more maintainable and easier for LLMs to contribute to. The new module encapsulates all security policy, permission checking, and sandboxing logic.

The `RexxInterpreter` class has been updated to import and delegate to the new security module, ensuring that the existing functionality remains unchanged.
